+++ app/libraries/Jentil/Setups/Translations/Breadcrumbs.php
@@ -26,9 +26,9 @@ final class Breadcrumbs extends AbstractTranslation
     {
         \load_theme_textdomain(
             $this->textDomain,
-            $this->app->utilities->fileSystem->vendorDir(
+            $this->app->utilities->fileSystem->dir(
                 'path',
-                '/grottopress/wordpress-breadcrumbs/src/lang'
+                '/dist/vendor/grotto-wp-breadcrumbs'
             )
         );
     }