+++ app/libraries/Jentil/Setups/Translations/Field.php
@@ -26,9 +26,9 @@ final class Field extends AbstractTranslation
     {
         \load_theme_textdomain(
             $this->textDomain,
-            $this->app->utilities->fileSystem->vendorDir(
+            $this->app->utilities->fileSystem->dir(
                 'path',
-                '/grottopress/wordpress-field/src/lang'
+                '/dist/vendor/grotto-wp-field'
             )
         );
     }