+++ app/libraries/Jentil/Setups/Translations/Posts.php
@@ -26,9 +26,9 @@ final class Posts extends AbstractTranslation
     {
         \load_theme_textdomain(
             $this->textDomain,
-            $this->app->utilities->fileSystem->vendorDir(
+            $this->app->utilities->fileSystem->dir(
                 'path',
-                '/grottopress/wordpress-posts/src/lang'
+                '/dist/vendor/grotto-wp-posts'
             )
         );
     }